Move file.
